Cambios Filtro Rango de Precios
Se ajusta el campo Price de tipo integer a Long. Se implementan los
cambios requeridos en el back y front para aplicar el nuevo filtro de la
búsqueda avanzada.

+++ mpcellphone.logic/src/main/java/co/edu/uniandes/csw/mpcellphone/api/IProductLogic.java
@@ -18,7 +18,7 @@ public interface IProductLogic {
     public List<ProductDTO> getByBrand(String brand);
     public List<ProductDTO> getByProviderName(String name);
     public List<ProductDTO> getByCity(String city);
-    public List<ProductDTO> getByPriceRange(Integer minPrice, Integer maxPrice);
+    public List<ProductDTO> getByPriceRange(Long minPrice, Long maxPrice);
     public List<ProductDTO> getByDiscount();
     public List<ProductDTO> getByCategory(String category);
     public List<ProductDTO> getCategories();
